fix(bph-catalog): move contributor type/roles to a client-safe module
The Turbopack build failed (dns/fs/child_process not resolvable) because
ContributorsEditor ('use client') value-imported BPH_CONTRIBUTOR_ROLES from
bph-catalog.ts, which imports mongodb/supabase — dragging server-only modules
into the browser bundle. tsc doesn't catch this (types erase; the bundler
boundary is the real check).

Move BphContributor + BPH_CONTRIBUTOR_ROLES into src/lib/bph-contributors.ts
(no server deps); bph-catalog re-exports them for server callers; client
components import from the new module.
